productspage

+++ src/test/java/StepDefinitions/MasterHooks.java
@@ -1,8 +1,12 @@
 package StepDefinitions;
 
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
+
 import Utils.DriverFactory;
 import io.cucumber.java.After;
 import io.cucumber.java.Before;
+import io.cucumber.java.Scenario;
 
 public class MasterHooks extends DriverFactory{
 	
@@ -18,12 +22,24 @@ public class MasterHooks extends DriverFactory{
 //		driver=getDriver();
 //	}
 	@After
-	public void tearDown()
+	public void tearDownScreenshotOnFailure(Scenario scenario)
 	{
-		if(driver!=null)
-		{
-			driver.manage().deleteAllCookies();
-			driver.quit();
+		try {
+			if(driver!=null &&scenario.isFailed())
+			{
+				scenario.attach(((TakesScreenshot) driver).getScreenshotAs(OutputType.BYTES),"image/png", null);
+				driver.manage().deleteAllCookies();
+				driver.quit();
+				driver=null;
+			}
+			if(driver!=null)
+			{
+				driver.manage().deleteAllCookies();
+				driver.quit();
+				driver=null;
+			}
+		} catch (Exception e) {
+			System.out.println("Method failed: tearDownScreenShotOnFailure, Exception: "+e.getMessage());;
 		}
 	}
